split editor and viewer to separate files.  link things up.
